Add rename-capable copy and move extension methods and interfaces
