+++ src/Extensions/IMoveFrom.cs
@@ -1,12 +1,11 @@
-﻿using System;
-using System.IO;
+﻿using System.IO;
 using System.Threading;
 using System.Threading.Tasks;
 
 namespace OwlCore.Storage;
 
 /// <summary>
-/// Provides a fast-path for the <see cref="ModifiableFolderExtensions.MoveFromAsync"/> extension method.
+/// Provides a fast-path for the <see cref="ModifiableFolderExtensions.CreateCopyOfAsync(IModifiableFolder, IFile, bool, CancellationToken)"/> extension method.
 /// </summary>
 /// <exception cref="FileNotFoundException">The item was not found in the provided folder.</exception>
 public interface IMoveFrom : IModifiableFolder
@@ -27,3 +26,30 @@ public interface IMoveFrom : IModifiableFolder
 /// A delegate that provides a fallback for the <see cref="IMoveFrom.MoveFromAsync"/> method.
 /// </summary>
 public delegate Task<IChildFile> MoveFromDelegate(IModifiableFolder modifiableFolder, IChildFile file, IModifiableFolder source, bool overwrite, CancellationToken cancellationToken);
+
+/// <summary>
+/// Provides a fast-path for the <see cref="ModifiableFolderExtensions.CreateCopyOfAsync(IModifiableFolder, IFile, bool, string, CancellationToken)"/> extension method.
+/// </summary>
+/// <remarks>
+/// This interface derives from <see cref="IMoveFrom"/> to enforce non-fallback compatibility with the non-rename overload.
+/// </remarks>
+/// <exception cref="FileNotFoundException">The item was not found in the provided folder.</exception>
+public interface IMoveRenamedFrom : IMoveFrom
+{
+    /// <summary>
+    /// Moves a storable item out of the provided folder, and into this folder. Returns the new item that resides in this folder.
+    /// </summary>
+    /// <param name="fileToMove">The file being moved into this folder.</param>
+    /// <param name="source">The folder that <paramref name="fileToMove"/> is being moved from.</param>
+    /// <param name="overwrite">If there is an existing destination file, <c>true</c> will overwrite it; otherwise <c>false</c> and the existing file is opened.</param>
+    /// <param name="newName">The new name of the created file.</param>
+    /// <param name="cancellationToken">A token that can be used to cancel the ongoing operation.</param>
+    /// <param name="fallback">The fallback to use if the provided <paramref name="fileToMove"/> isn't supported.</param>
+    /// <returns>The newly created (or opened if existing) file.</returns>
+    Task<IChildFile> MoveFromAsync(IChildFile fileToMove, IModifiableFolder source, bool overwrite, string newName, CancellationToken cancellationToken, MoveRenamedFromDelegate fallback);
+}
+
+/// <summary>
+/// A delegate that provides a fallback for the <see cref="IMoveFrom.MoveFromAsync"/> method.
+/// </summary>
+public delegate Task<IChildFile> MoveRenamedFromDelegate(IModifiableFolder modifiableFolder, IChildFile file, IModifiableFolder source, bool overwrite, string newName, CancellationToken cancellationToken);
